Fix ssh-dss signature verification with OpenSSL 3+
The PR #34 shows a failling test on systems using OpenSSL 3.  By
tracking the code execution, we can see that after computing the DER
signature, it is decoded to ASN.1, and that both values in the ASN.1
message are handled separately as `sign_r` and `sign_s`, which are
padded to 20 chars by "\0".

On Debian 11 with OpenSSL 1.1.1n and Ruby 2.7.4p191, `sign_r` and
`sign_s` are 20 bytes long, while on Ubuntu 22 with OpenSSL 3.0.2 and
Ruby 3.0.2p107 they are 28 bytes long.

This is not an issue when building `signature_h` where we concatenate
`sign_r` and `sign_s` because right padding a 28 chars long string to 20
chars does not change it; but for verification we read `sign_r` and
`sign_s` from fixed position in `signature_h[:'signature blob']`.  As a
consquence, the read `sign_r` is truncated and the read `sign_s` is
mangled with the end of `sign_r` and the beginnig of the expected
`sign_s`.  Of course, the signature verification then fails.

Right-padding to 28 chars instead of 28 and adjusing the offsets
accordingly seems to verify as expected on all Ruby/OpenSSL versions.